split individual file uploaders
